(BSR)[PRO] feat: split individual booking and collective
