+++ pro/src/components/Bookings/Bookings.tsx
@@ -27,8 +27,8 @@ import { selectCurrentOffererId } from '@/commons/store/offerer/selectors'
 import { isEqual } from '@/commons/utils/isEqual'
 import { CollectiveBudgetCallout } from '@/components/CollectiveBudgetInformation/CollectiveBudgetCallout'
 import { NoData } from '@/components/NoData/NoData'
-import { ChoosePreFiltersMessage } from '@/pages/Bookings/ChoosePreFiltersMessage/ChoosePreFiltersMessage'
-import { NoBookingsForPreFiltersMessage } from '@/pages/Bookings/NoBookingsForPreFiltersMessage/NoBookingsForPreFiltersMessage'
+import { ChoosePreFiltersMessage } from '@/pages/IndividualBookings/ChoosePreFiltersMessage/ChoosePreFiltersMessage'
+import { NoBookingsForPreFiltersMessage } from '@/pages/IndividualBookings/NoBookingsForPreFiltersMessage/NoBookingsForPreFiltersMessage'
 import { Spinner } from '@/ui-kit/Spinner/Spinner'
 
 import { BookingsRecapTable } from './BookingsRecapTable/BookingsRecapTable'
