+++ pro/src/pages/IndividualBookings/adapters/getUserHasIndividualBookingsAdapter.ts
